# significantly speed up parsing of easyconfig files by only extracting comments from an easyconfig file when they're actually needed